requirements.txt
### THE BELOW VERSIONS ARE CURRENTLY WORKING IN THE PRODUCTION ENVIRONMENT ###

ciscoconfparse==1.6.40
Flask==2.0.3
genie==22.2
graphviz==0.19.1
httpx==0.22.0
ipfabric==5.0.12
netaddr==0.7.19
nornir==3.2.0
nornir-napalm==0.2.0
nornir-scrapli==2022.1.30
panapi==0.0.1
pyats==22.2
python-decouple==3.6
PyYAML==6.0
regex==2022.7.25
requests==2.27.1
tqdm==4.63.1
urllib3==1.26.9
waitress==2.1.1
Werkzeug==2.0.3
xmltodict==0.12.0
yamllint==1.26.3

# To install: pip3 install
# To install all libraries listed here: pip3 install -r requirements.txt
# I suggest you create a virtua environment prior to this as shown in the documentation folder.
### IN ADDITION TO THE ABOVE PYTHON LIBRARIES, THE BELOW MUST ALSO BE INSTALLED VIA APT-GET INSTALL OR APT-GET
# python3-pip
# python3-venv
# graphviz
# nginx
# node js
# npm